Massive health bar logic refactor + health bar instancing prototype
