+++ src/main/java/io/enderdev/endermodpacktweaks/features/healthbar/HealthBarHandler.java
@@ -1,0 +1,261 @@
+package io.enderdev.endermodpacktweaks.features.healthbar;
+
+import io.enderdev.endermodpacktweaks.config.CfgFeatures;
+import io.enderdev.endermodpacktweaks.mixin.minecraft.WorldClientAccessor;
+import io.enderdev.endermodpacktweaks.utils.EmtConfigHandler;
+import io.enderdev.endermodpacktweaks.utils.EmtConfigParser;
+import io.enderdev.endermodpacktweaks.utils.EmtRender;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.renderer.culling.Frustum;
+import net.minecraft.client.settings.KeyBinding;
+import net.minecraft.entity.*;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.util.math.AxisAlignedBB;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.RayTraceResult;
+import net.minecraft.util.math.Vec3d;
+import net.minecraft.world.World;
+import net.minecraftforge.client.event.RenderWorldLastEvent;
+import net.minecraftforge.fml.client.registry.ClientRegistry;
+import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
+import net.minecraftforge.fml.common.gameevent.InputEvent;
+import org.lwjgl.opengl.GL11;
+import org.lwjgl.opengl.GL20;
+import org.lwjgl.util.vector.Vector3f;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Stack;
+
+public class HealthBarHandler {
+    public static boolean ENABLE_INSTANCING = false;
+    private BarInstancingRenderer backgroundRenderer = null;
+
+    public final EmtConfigHandler<EmtConfigParser.ConfigItem> whitelist = new EmtConfigHandler<>(
+            CfgFeatures.MOB_HEALTH_BAR.onlyRenderWithEquipment,
+            EmtConfigParser.ConfigItem::new
+    );
+
+    public final EmtConfigHandler<EmtConfigParser.ConfigItemWithFloat> rangeModifiers = new EmtConfigHandler<>(
+            CfgFeatures.MOB_HEALTH_BAR.distanceMultipliers,
+            EmtConfigParser.ConfigItemWithFloat::new
+    );
+
+    private final KeyBinding key;
+    private boolean down;
+
+    private boolean shouldRender = true;
+
+    public HealthBarHandler() {
+        key = new KeyBinding("keybind.endermodpacktweaks.toggle", 0, "key.categories.misc");
+        ClientRegistry.registerKeyBinding(key);
+    }
+
+    @SubscribeEvent
+    public void onKeyInput(InputEvent.KeyInputEvent event) {
+        Minecraft mc = Minecraft.getMinecraft();
+        boolean wasDown = down;
+        down = key.isKeyDown();
+        if (mc.inGameHasFocus && down && !wasDown) {
+            shouldRender = !shouldRender;
+        }
+    }
+
+    private final Frustum frustum = new Frustum();
+
+    @SubscribeEvent
+    public void onRenderWorldLast(RenderWorldLastEvent event) {
+
+        // early escape
+        Minecraft mc = Minecraft.getMinecraft();
+        if ((!CfgFeatures.MOB_HEALTH_BAR.renderInF1 && !Minecraft.isGuiEnabled()) || !shouldRender) return;
+        if (CfgFeatures.MOB_HEALTH_BAR.onlyRenderWithEquipment.length != 0 && !whitelist.equipped(mc.player)) return;
+        Entity cameraEntity = mc.getRenderViewEntity();
+        if (cameraEntity == null || !cameraEntity.isEntityAlive()) return;
+
+        BlockPos cameraBlockPos = cameraEntity.getPosition();
+        Vector3f cameraPos = EmtRender.getCameraPos();
+        float partialTicks = (float) EmtRender.getPartialTick();
+
+        frustum.setPosition(cameraPos.x, cameraPos.y, cameraPos.z);
+
+        // init instancing renderer
+        if (ENABLE_INSTANCING) {
+            if (backgroundRenderer == null)
+                backgroundRenderer = (new BarInstancingRenderer(100)).init();
+        }
+
+        // render health bar for these entities
+        List<EntityLivingBase> entities = new ArrayList<>();
+
+        // collect entities
+        if (CfgFeatures.MOB_HEALTH_BAR.showOnlyFocused) {
+            Entity focused = getEntityLookedAt(mc.player);
+            if (focused instanceof EntityLivingBase && focused.isEntityAlive()) {
+                collectHealthBarEntities(entities, (EntityLivingBase) focused, cameraEntity);
+            }
+        } else {
+            List<Entity> filteredEntities = new ArrayList<>();
+            for (Entity entity : ((WorldClientAccessor) mc.world).getEntityList()) {
+                if (entity instanceof EntityLivingBase
+                        && entity != mc.player
+                        && entity.isInRangeToRender3d(cameraBlockPos.getX(), cameraBlockPos.getY(), cameraBlockPos.getZ())
+                        && (entity.ignoreFrustumCheck || frustum.isBoundingBoxInFrustum(entity.getEntityBoundingBox()))
+                        && entity.isEntityAlive()
+                        && entity.getRecursivePassengers().isEmpty()) {
+                    collectHealthBarEntities(entities, (EntityLivingBase) entity, cameraEntity);
+                }
+            }
+        }
+
+        // optifine compat: disable shader program
+        int oldProgram = GL11.glGetInteger(GL20.GL_CURRENT_PROGRAM);
+        if (oldProgram != 0) GL20.glUseProgram(0);
+        for (EntityLivingBase entity : entities) {
+            HealthBarRenderHelper.renderHealthBar(entity, partialTicks, false);
+        }
+        if (oldProgram != 0) GL20.glUseProgram(oldProgram);
+
+//        if (backgroundRenderer != null) {
+//            int entityListLength = Math.min(backgroundRenderer.getMaxInstance(), filteredEntities.size());
+//            float[] instanceData = new float[backgroundRenderer.getMaxInstance() * 3];
+//            for (int i = 0; i < entityListLength; i++) {
+//                Entity entity = filteredEntities.get(i);
+//                instanceData[i * 3] = (float) (entity.lastTickPosX + (entity.posX - entity.lastTickPosX) * partialTicks);
+//                instanceData[i * 3 + 1] = (float) (entity.lastTickPosY + (entity.posY - entity.lastTickPosY) * partialTicks) + 1.6f;
+//                instanceData[i * 3 + 2] = (float) (entity.lastTickPosZ + (entity.posZ - entity.lastTickPosZ) * partialTicks);
+//            }
+//
+//            backgroundRenderer.getMesh().setInstancePrimCount(entityListLength);
+//            backgroundRenderer.getMesh().updateInstanceDataByBufferSubData(instanceData);
+//
+//            GlStateManager.enableBlend();
+//
+//            backgroundRenderer.getShaderProgram().use();
+//            backgroundRenderer.getShaderProgram().setUniform("modelView", EmtRender.getModelViewMatrix());
+//            backgroundRenderer.getShaderProgram().setUniform("projection", EmtRender.getProjectionMatrix());
+//            backgroundRenderer.getShaderProgram().setUniform("camPos", cameraPos.x, cameraPos.y, cameraPos.z);
+//            backgroundRenderer.getShaderProgram().unuse();
+//
+//            backgroundRenderer.render();
+//        }
+    }
+
+    private void collectHealthBarEntities(List<EntityLivingBase> entities, EntityLivingBase entity, Entity viewPoint) {
+        Stack<EntityLivingBase> ridingStack = new Stack<>();
+        ridingStack.push(entity);
+
+        while (entity.getRidingEntity() != null && entity.getRidingEntity() instanceof EntityLivingBase) {
+            entity = (EntityLivingBase) entity.getRidingEntity();
+            ridingStack.push(entity);
+        }
+
+        Minecraft mc = Minecraft.getMinecraft();
+
+        while (!ridingStack.isEmpty()) {
+            entity = ridingStack.pop();
+            boolean boss = !entity.isNonBoss();
+
+            String entityID = EntityList.getEntityString(entity);
+            if (Arrays.asList(CfgFeatures.MOB_HEALTH_BAR.mobBlacklist).contains(entityID)) {
+                continue;
+            }
+
+            processing:
+            {
+                float maxDistance = CfgFeatures.MOB_HEALTH_BAR.maxDistance;
+                if (CfgFeatures.MOB_HEALTH_BAR.distanceMultipliers.length != 0 && rangeModifiers.equipped(mc.player)) {
+                    EmtConfigParser.ConfigItemWithFloat modifier = (EmtConfigParser.ConfigItemWithFloat) rangeModifiers.getEquipped(mc.player);
+                    maxDistance *= modifier != null ? modifier.value() : 1F;
+                }
+                float distance = entity.getDistance(viewPoint);
+                if (distance > maxDistance || !entity.canEntityBeSeen(viewPoint) || entity.isInvisible()) {
+                    break processing;
+                }
+                if (!CfgFeatures.MOB_HEALTH_BAR.showOnBosses && boss) {
+                    break processing;
+                }
+                if (!CfgFeatures.MOB_HEALTH_BAR.showOnPlayers && entity instanceof EntityPlayer) {
+                    break processing;
+                }
+
+                if (entity.getMaxHealth() <= 0) {
+                    break processing;
+                }
+
+                entities.add(entity);
+            }
+        }
+    }
+
+    private Entity getEntityLookedAt(Entity e) {
+        Entity foundEntity = null;
+
+        double finalDistance = CfgFeatures.MOB_HEALTH_BAR.maxDistance;
+        if (CfgFeatures.MOB_HEALTH_BAR.distanceMultipliers.length != 0 && rangeModifiers.equipped((EntityPlayer) e)) {
+            EmtConfigParser.ConfigItemWithFloat modifier = (EmtConfigParser.ConfigItemWithFloat) rangeModifiers.getEquipped((EntityPlayer) e);
+            finalDistance *= modifier != null ? modifier.value() : 1F;
+        }
+        double distance = finalDistance;
+        RayTraceResult pos = rayCast(e, finalDistance);
+        Vec3d positionVector = e.getPositionVector();
+        if (e instanceof EntityPlayer) {
+            positionVector = positionVector.add(0, e.getEyeHeight(), 0);
+        }
+
+        if (pos != null) {
+            distance = pos.hitVec.distanceTo(positionVector);
+        }
+
+        Vec3d lookVector = e.getLookVec();
+        Vec3d reachVector = positionVector.add(lookVector.x * finalDistance, lookVector.y * finalDistance, lookVector.z * finalDistance);
+
+        Entity lookedEntity = null;
+        List<Entity> entitiesInBoundingBox = e.getEntityWorld().getEntitiesWithinAABBExcludingEntity(e, e.getEntityBoundingBox().grow(lookVector.x * finalDistance, lookVector.y * finalDistance, lookVector.z * finalDistance).expand(1F, 1F, 1F));
+        double minDistance = distance;
+
+        for (Entity entity : entitiesInBoundingBox) {
+            if (entity.canBeCollidedWith()) {
+                float collisionBorderSize = entity.getCollisionBorderSize();
+                AxisAlignedBB hitbox = entity.getEntityBoundingBox().expand(collisionBorderSize, collisionBorderSize, collisionBorderSize);
+                RayTraceResult interceptPosition = hitbox.calculateIntercept(positionVector, reachVector);
+
+                if (hitbox.contains(positionVector)) {
+                    if (0.0D < minDistance || minDistance == 0.0D) {
+                        lookedEntity = entity;
+                        minDistance = 0.0D;
+                    }
+                } else if (interceptPosition != null) {
+                    double distanceToEntity = positionVector.distanceTo(interceptPosition.hitVec);
+
+                    if (distanceToEntity < minDistance || minDistance == 0.0D) {
+                        lookedEntity = entity;
+                        minDistance = distanceToEntity;
+                    }
+                }
+            }
+
+            if (lookedEntity != null && (minDistance < distance || pos == null)) {
+                foundEntity = lookedEntity;
+            }
+        }
+
+        return foundEntity;
+    }
+
+    private static RayTraceResult rayCast(Entity e, double len) {
+        Vec3d vec = new Vec3d(e.posX, e.posY, e.posZ);
+        if (e instanceof EntityPlayer) {
+            vec = vec.add(new Vec3d(0, e.getEyeHeight(), 0));
+        }
+
+        Vec3d look = e.getLookVec();
+        return rayCast(e.getEntityWorld(), vec, look, len);
+    }
+
+    private static RayTraceResult rayCast(World world, Vec3d origin, Vec3d ray, double len) {
+        Vec3d end = origin.add(ray.normalize().scale(len));
+        return world.rayTraceBlocks(origin, end);
+    }
+}
